+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - T0152 on 2025-09-02
 - T0155 on 2025-09-02
 
+- T0167 on 2025-09-02
+- T0168 on 2025-09-02
 ## [v0.2.0] - 2025-08-31
 
 ### Added