+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.1] - 2025-09-10
+
 ### Added
 - Radar chart top 10 models.
 - Zettelkatalog benchmark.
@@ -27,9 +29,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - T0151 on 2025-09-02
 - T0152 on 2025-09-02
 - T0155 on 2025-09-02
-
 - T0167 on 2025-09-02
 - T0168 on 2025-09-02
+
 ## [v0.2.0] - 2025-08-31
 
 ### Added
@@ -54,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changelog.
 
-[Unreleased]: https://github.com/RISE-UNIBAS/humanities_data_benchmark/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/RISE-UNIBAS/humanities_data_benchmark/compare/v0.2.1...HEAD
 [v0.1.0]: https://github.com/RISE-UNIBAS/humanities_data_benchmark/releases/tag/v0.1.0
 [v0.2.0]: https://github.com/RISE-UNIBAS/humanities_data_benchmark/releases/tag/v0.2.0
+[v0.2.1]: https://github.com/RISE-UNIBAS/humanities_data_benchmark/releases/tag/v0.2.1